refactor(solid): combine into one file
