--- packages/use-wallet-solid/src/WalletProvider.tsx
@@ -1,31 +1,0 @@
-import { WalletManager } from '@txnlab/use-wallet'
-import { createContext, useContext, JSX, onMount } from 'solid-js'
-
-interface WalletProviderProps {
-  manager: WalletManager
-  children: JSX.Element
-}
-
-const WalletContext = createContext<() => WalletManager>()
-
-export const WalletProvider = (props: WalletProviderProps) => {
-  const store = () => props.manager
-
-  onMount(async () => {
-    try {
-      await props.manager.resumeSessions()
-    } catch (error) {
-      console.error('Error resuming sessions:', error)
-    }
-  })
-
-  return <WalletContext.Provider value={store}>{props.children}</WalletContext.Provider>
-}
-
-export const useWalletManager = (): WalletManager => {
-  const manager = useContext(WalletContext)
-  if (!manager) {
-    throw new Error('useWalletManager must be used within a WalletProvider')
-  }
-  return manager()
-}
